Split FieldConverter and SchemaConverter from ArrowConverter
