+++ java/ingest/ingest-runner/src/main/java/sleeper/ingest/impl/recordbatch/arrow/ArrowRecordBatchBase.java
@@ -47,7 +47,7 @@ import java.util.Objects;
 import java.util.UUID;
 
 import static java.util.Objects.requireNonNull;
-import static sleeper.arrow.ArrowConverter.convertSleeperSchemaToArrowSchema;
+import static sleeper.arrow.schema.SchemaConverter.convertSleeperSchemaToArrowSchema;
 
 /**
  * This class stores data in Arrow format, in-memory where possible and then spilling to disk once the memory is full.
